feat: 문진 결과 리스트 화면 추가
