+++ src/App.tsx
@@ -1,6 +1,16 @@
 import { Routes, Route, Navigate } from "react-router-dom";
 import { useEffect } from "react";
-import { Diagnosis, DiagnosisList, Information, MainPage, ResultPage, SymptomPage, SymptomTypePage, HomeLoading } from "./pages";
+import {
+  Diagnosis,
+  MyDiagnosis,
+  Information,
+  MainPage,
+  ResultPage,
+  SymptomPage,
+  SymptomTypePage,
+  HomeLoading,
+  DiagnosisList,
+} from "./pages";
 import { useAppSelector } from "./state";
 import styled from "styled-components";
 
@@ -49,8 +59,9 @@ function App() {
     <Main>
       <Container>
         <Routes>
-          <Route path="/" element={authenticated ? <DiagnosisList /> : <MainPage />} />
+          <Route path="/" element={authenticated ? <MyDiagnosis /> : <MainPage />} />
           <Route path="/info" element={<Information />} />
+          <Route path="/diagnosis-list" element={<DiagnosisList />} />
           <Route path="/result" element={<ResultPage />} />
           <Route path="/diagnosis" element={<Diagnosis />} />
           <Route path="/symptom" element={<SymptomPage />} />
