+++ src/components/diagnosisCard/index.style.tsx
