+++ src/pages/diagnosisList/index.tsx
@@ -1,74 +1,76 @@
-import { useLayoutEffect, useState } from "react";
-import RoundButton from "src/components/roundButton";
+import { useEffect } from "react";
+import { useLocation, useNavigate } from "react-router-dom";
+import { HeadacheDiagnose } from "src/api/diagnose";
+import DiagnosisCard from "src/components/diagnosisCard";
 import MainHeader from "src/components/mainHeader";
-import ListComponent from "./listComponent";
-import theme from "src/lib/theme";
-import { useNavigate } from "react-router-dom";
 import { IDiagnosisList } from "src/interfaces/component";
-import EmptyPage from "./emptyList";
-import { useAppSelector, useAppDispatch } from "src/state";
-import { DELETE_TOKEN } from "src/state/authSlice";
-import { Container, Title, DescriptionBox, List, ButtonBackground } from "./index.style";
-import { Diagnosis } from "src/api/diagnosis";
+import { IDiagnosisResultList } from "src/interfaces/diagnosticResult";
+import { Container, DescriptionText, List, Title } from "./index.style";
 
 const DiagnosisList = () => {
+  const { state } = useLocation() as { state: IDiagnosisResultList };
   const navigate = useNavigate();
-  const [diagnosisList, setDiagnosisList] = useState([] as IDiagnosisList[]);
-  const [name, setName] = useState("");
-  const [loading, setLoading] = useState(true);
-  const { accessToken } = useAppSelector((state) => state.auth);
-  const dispatch = useAppDispatch();
 
-  useLayoutEffect(() => {
-    const getList = async () => {
-      try {
-        const res = await Diagnosis.getDiagnosis(accessToken);
-
-        setDiagnosisList(res.diagnosis.reverse());
-        setName(res.nickname);
-        setLoading(false);
-      } catch (error) {
-        dispatch(DELETE_TOKEN);
-        alert("감별진단 목록 로딩 실패, 다시 시도해주세요");
-      }
-    };
-
-    getList();
+  useEffect(() => {
+    if (!state) navigate("/");
   }, []);
+
+  const handleNavigate = async (diag: IDiagnosisList) => {
+    const diagnosisResult = await HeadacheDiagnose.postResultDetail(Number(diag.record.diagnosis_id));
+
+    navigate("/result", {
+      state: {
+        type: "result",
+        diagnostic_result: diagnosisResult.diagnostic_result,
+      },
+    });
+  };
 
   return (
     <Container>
       <MainHeader />
-      {loading || (
+      <Title>
+        <span className="highlight">000님</span>의 예상 질환이에요
+      </Title>
+      <DescriptionText style={{ marginTop: "0.8rem" }}>
+        <span className="highlight">예상질환</span>을 클릭하면 자세한 설명을 확인할 수 있어요!
+      </DescriptionText>
+      {state && state.dataList.results.likely && (
         <>
-          {diagnosisList.length === 0 ? (
-            <EmptyPage />
-          ) : (
-            <>
-              <Title>
-                <span className="highlight">{name}님</span>
-                이 저장한
-                <br /> 증상 감별 내역이에요
-              </Title>
-              <DescriptionBox>
-                <span className="highlight">{diagnosisList.length}개</span>의 증상 감별 내역
-              </DescriptionBox>
-              <List>
-                {diagnosisList.map((diag, idx) => (
-                  <ListComponent key={idx} diagnosis={diag} />
-                ))}
-              </List>
-            </>
-          )}
+          <DescriptionText style={{ marginTop: "2.4rem" }}>
+            <span className="highlight">가장 가능성 높은 질환</span>
+          </DescriptionText>
+          <List>
+            {state.dataList.results.likely.map((diag, idx) => (
+              <DiagnosisCard key={idx} diagnosis={diag} handleNavigate={() => handleNavigate(diag)} />
+            ))}
+          </List>
         </>
       )}
-      <ButtonBackground>
-        <section className="button-box" onClick={() => navigate("/symptom-type", { state: "list" })}>
-          <RoundButton outline="none" backgroundColor={theme.color.green} color={theme.color.grey_900}>
-            빠른 증상 감별 시작하기
-          </RoundButton>
-        </section>
-      </ButtonBackground>
+      {state && state.dataList.results.predicted && (
+        <>
+          <DescriptionText style={{ marginTop: "2.4rem" }}>
+            <span className="highlight">예상질환이에요</span>
+          </DescriptionText>
+          <List>
+            {state.dataList.results.predicted.map((diag, idx) => (
+              <DiagnosisCard key={idx} diagnosis={diag} handleNavigate={() => handleNavigate(diag)} />
+            ))}
+          </List>
+        </>
+      )}
+      {state && state.dataList.results.suspicious && (
+        <>
+          <DescriptionText style={{ marginTop: "2.4rem" }}>
+            <span className="highlight">다음 질환도 의심되어요!</span>
+          </DescriptionText>
+          <List>
+            {state.dataList.results.suspicious.map((diag, idx) => (
+              <DiagnosisCard key={idx} diagnosis={diag} handleNavigate={() => handleNavigate(diag)} />
+            ))}
+          </List>
+        </>
+      )}
     </Container>
   );
 };
