+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import DiagnosisList from "./diagnosisList";
+import MyDiagnosis from "./myDiagnosis";
 import Diagnosis from "./diagnosis";
 import Information from "./information";
 import MainPage from "./main";
@@ -6,5 +6,6 @@ import ResultPage from "./result";
 import SymptomTypePage from "./symptomType";
 import SymptomPage from "./symptom";
 import HomeLoading from "./homeLoading";
+import DiagnosisList from "./diagnosisList";
 
-export { Diagnosis, DiagnosisList, Information, MainPage, ResultPage, SymptomPage, SymptomTypePage, HomeLoading };
+export { Diagnosis, MyDiagnosis, Information, MainPage, ResultPage, SymptomPage, SymptomTypePage, HomeLoading, DiagnosisList };
