+++ src/pages/myDiagnosis/emptyList/index.style.tsx
