+++ src/pages/myDiagnosis/emptyList/index.tsx
